Small code refactor
