+++ src/main/kotlin/com/nexsabre/hardwarereservationtool/cmd/data/ErrorData.kt
@@ -1,0 +1,3 @@
+package com.nexsabre.hardwarereservationtool.cmd.data
+
+data class ErrorData(val error: String, val message: String)
